Move IVR under dialplan.
The IVR had it' own context, but is now moved to the dialplan context
as a modal window.
The problem is that the IVR and dialplan have a dependency on each
other and with the split windows made it easy to delete an item
the other needed. That situation is not fixed with this patch, but
the groundwork for a fix is layed.